Prefab and script for BattlePopupSelectedCharacter created
- Created own folder for SelectedCharacters scripts and moved BattlePopupCharacterSlotController.cs there.
- Replaced the GalleryCharacter prefabs in SelectedCharacters prefab with BattlePopupSelectedCharacter prefabs.
- Created namespace matching the folder path for the two script files. BattlePopupSelectedCharacter.cs and BattlePopupCharacterSlotController.cs
