+++ Assets/MenuUi/Prefabs/UI Components/Battle Popup/BattlePopupSelectedCharacter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 989bf4b949354394bbb3cfa7ebcfda90
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
