+++ Assets/MenuUi/Scripts/Lobby/SelectedCharacters.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 76cd81c14fc6d954f815c1d58c37f1b6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
